% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{aaaaa}
\alias{aaaaa}
\title{aaaaa: A wAy to pAinlessly switch between different power LAw formulAtions}
\usage{
aaaaa(intercept, from = c("Ad", "ad", "AD", "aD"), to = c("Ad", "ad", "AD",
  "aD"), slope, n)
}
\arguments{
\item{intercept}{Intercept parameter to be converted.}

\item{from}{Kind of the input intercept parameter.}

\item{to}{Desired kind for the ouput intercept parameter.}

\item{slope}{Slope parameter.}

\item{n}{Number of individuals per sampling unit.}
}
\description{
\code{aaaaa} was designed to avoid headaches that are likely to occur when
working with different formulations of the binomial power law analysis.
}
\details{
The binomial power law can be expressed as: \eqn{s_y^2 = (intercept)(s_{bin}^2)^b}.
But different forms of (intercept) are possible depending on the formulation of the
binomial power law.
\tabular{ccccc}{
      \tab Ad         \tab ad      \tab AD         \tab aD      \cr
   Ad \tab 1          \tab n^b     \tab n^{2(b-1)} \tab n^{b-2} \cr
   ad \tab n^{-b}     \tab 1       \tab n^{b-2}    \tab n^{-2}  \cr
   AD \tab n^{2(1-b)} \tab n^{2-b} \tab 1          \tab n^{-b}  \cr
   aD \tab n^{2-b}    \tab n^2     \tab n^b        \tab 1       \cr
}
}
\examples{

aaaaa(from = , to = , n = , b = )
aaaaa(to = , data = <powerLaw object>)

}
