% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{aggIndices}
\alias{aggIndices}
\title{Patchiness/Aggregation/Clustering indices}
\usage{
aggIndices(data, method = "all")
}
\arguments{
\item{data}{The data set you want to explore.}

\item{method}{Method implemented.}
}
\description{
TO DO
}
\details{
Different indices are available, and depend on the nature of the data set.
Only indices for \code{Count} and \code{Incidence} data are currently
implemented.
}
