% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aphids}
\alias{aphids}
\title{Aphid counts}
\format{A data frame with 63 rows and 3 variables:
\tabular{lll}{
    [, 1:2] \tab x,y \tab Spatial coordonates. \cr
    [, 3]   \tab d   \tab Counts. \cr
}}
\source{
Perry JN, Winder L, Holland JM, Alston RD. 1999. Red-blue plots for
    detecting clusters in count data. Ecology Letters 2, 106-13.
    \href{http://dx.doi.org/10.1046/j.1461-0248.1999.22057.x}{10.1046/j.1461-0248.1999.22057.x}
}
\usage{
aphids
}
\description{
TODO.
}
\keyword{datasets}
