% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropods}
\alias{arthropods}
\title{Counts of arthropods.}
\format{
A data frame with 378 rows and 4 variables:
\tabular{lll}{
    [, 1:2] \tab x,y   \tab Grid spatial coordinates. \cr
    [, 3:4] \tab xm,ym \tab Metric spatial coordonates. \cr
    [, 5]   \tab t     \tab Sampling date. 1: 7 Jun, 2: 14 Jun, 3: 21 Jun, 4:
                            28 Jun, 5: 5 Jul, 6: 12 Jul 1996. \cr
    [, 6]   \tab i     \tab Counts of arthropods. \cr
}
}
\source{
Holland JM, Winder L, Perry JN. 1999. Arthropod prey of farmland
    birds: Their spatial distribution within a sprayed field with and without
    buffer zones. Aspects of Applied Biology 54: 53-60.
}
\usage{
arthropods
}
\description{
A sampling unit was made of a pitfall to collect arthropods in a field of
organic winter wheat, near Wimborne, Dorset, UK in 1996. The sampling units
were located on a 9 x 7 rectangular grid at intervals of 30 m. There were
six sampling dates.
}
\keyword{datasets}
