% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropods_counts}
\alias{arthropods_counts}
\title{Arthropods counts}
\format{A data frame with 378 rows and 4 variables:
\tabular{lll}{
    [, 1]   \tab set \tab Grid id. \cr
    [, 2:3] \tab x,y \tab Spatial coordonates. \cr
    [, 4]   \tab d   \tab Counts. \cr
}}
\source{
Holland JM, Winder L, Perry JN. 1999. Arthropod prey of farmland
    birds: Their spatial distribution within a sprayed field with and without
    buffer zones. Aspects of Applied Biology 54: 53-60.
}
\usage{
arthropods_counts
}
\description{
TODO.
}
\keyword{datasets}
