% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{as.array}
\alias{as.array}
\alias{as.array}
\alias{as.array.intensity}
\title{Test}
\usage{
\method{as.array}{intensity}(x, dim, ...)
}
\arguments{
\item{x}{An intensity object.}

\item{dim}{Name or ID of the dimensions.}
}
\description{
Test
}
\examples{

myData <- incidence(Cochran1936)
as.array(myData, dim = c(24, 60, 2))

}
