% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.intensity}
\title{Coerce to a data frame}
\usage{
\method{as.data.frame}{intensity}(x, ..., members = c("labels", "space",
  "time", "obs"))
}
\arguments{
\item{x}{An \code{intensity} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{members}{Slots of \code{x} to use to build the data frame.}
}
\value{
A data frame.
}
\description{
Functions to coerce an \code{intensity} object to data.frame class.
}
\examples{
myData <- incidence(Cochran1936)
head(as.data.frame(myData))
head(as.data.frame(myData, fields = c("space", "obs")))

}
\seealso{
\code{\link{count}} \code{\link{incidence}} \code{\link{severity}}
}
