% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{as.incidence}
\alias{as.incidence}
\alias{as.incidence}
\alias{as.incidence.count}
\alias{as.incidence}
\alias{as.incidence.severity}
\title{as.incidence & is.incidence}
\usage{
as.incidence(object, ...)

\method{as.incidence}{count}(object, ...)

\method{as.incidence}{severity}(object, ...)
}
\arguments{
\item{object}{Any R object.}
}
\value{
\code{as.incidence} returns an \code{incidence} object if it is possible.

\code{is.incidence} returns \code{TRUE} if its argument is an \code{incidence} object
(that is, has \code{incidence} amongst its classes) and \code{FALSE} otherwise.
}
\description{
Functions to check if an object is an \code{\link{incidence}} object, or coerce it if possible.
}
\details{
\code{as.incidence} is a generic function with different methods. In this package, only
methods for \code{\link{count}} and \code{\link{severity}} objects are implemented.

\code{is.incidence} is basically a wrapper.
}
\seealso{
\code{\link{incidence}}
}
