% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{BetaBinomial}
\alias{BetaBinomial}
\alias{dbetabinom}
\alias{pbetabinom}
\alias{qbetabinom}
\alias{rbetabinom}
\title{The Beta-Binomial Distribution}
\usage{
dbetabinom(x, size, prob, theta, shape1, shape2, log = FALSE)

pbetabinom(q, size, prob, theta, shape1, shape2, lower.tail = TRUE,
  log.p = FALSE)

qbetabinom(p, size, prob, theta, shape1, shape2, lower.tail = TRUE,
  log.p = FALSE)

rbetabinom(n, size, prob, theta, shape1, shape2)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{size}{Number of trials.}

\item{prob}{Probability of success on each trial.}

\item{theta}{Aggregation parameter (theta = 1 / (shape1 + shape2)).}

\item{shape1, shape2}{Shape parameters.}

\item{log, log.p}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{[Not yet implemented].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dbetabinom} gives the density, \code{pbetabinom} gives the distribution
function, \code{qbetabinom} gives the quantile function and \code{rbetabinom}
generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the beta-binomial distribution with parameters \code{size}, \code{prob},
\code{theta}, \code{shape1}, \code{shape2}. It corresponds to as an
overdispersed binomial distribution.
}
\details{
Be careful, in this implementation, theta = 1 / (shape1 + shape2).
\code{prob} and \code{theta}, or \code{shape1} and \code{shape2} must be specified.
if \code{theta} = 0, use *binom family instead.
}
\examples{
# Compute P(25 < X < 50) for X Beta-Binomial(100, 0.5, 0.35)
sum(dbetabinom(25:50, size = 100, prob = 0.5, theta = 0.35))

# When theta -> 0, dbetabinom -> dbinom
sum(dbetabinom(25:50, size = 100, prob = 0.5, theta = 1e-7))
sum(dbetabinom(25:50, size = 100, shape1 = 1e7, shape2 = 1e7))
sum(dbinom(25:50, size = 100, prob = 0.5))

n <- 15
q <- 0:n
p1 <- dbinom(q, size = n, prob = 0.33)
p2 <- dbetabinom(q, size = n, prob = 0.33, theta = 0.22)
res <- rbind(p1, p2)
dimnames(res) <- list(c("Binomial", "Beta-binomial"), q)
barplot(res, beside=TRUE, legend.text = TRUE, ylab = "Frequency")

}
