% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citrus_ctv}
\alias{citrus_ctv}
\title{Incidence of citrus tristeza virus (CTV) disease in three fields.}
\format{There are three data sets:
\itemize{
    \item \code{IVI3and4}: A data frame with 864 rows and 5 variables.
    \item \code{IVI6and7}: A data frame with 648 rows and 5 variables.
    \item \code{El_Realengo}: A data frame with 2000 rows and 5 variables.
}

The structure is the same for all the data sets:

\tabular{rll}{
    [, 1:2] \tab x,y \tab Spatial coordinates (in a grid). \cr
    [, 3]   \tab t   \tab Year of disease assessments. \cr
    [, 4]   \tab r   \tab Disease incidence. 0: Healthy, 1: Diseased. \cr
    [, 5]   \tab n   \tab Sampling unit size. n = 1 means that the sampling unit size is the plant. \cr
}}
\source{
Gottwald TR, Cambra M, Moreno P, Camarasa E, Piquer J. 1996. Spatial
    and temporal analyses of citrus tristeza virus in eastern Spain.
    Phytopathology 86, 45–55.

Gibson GJ. 1997. Investigating mechanisms of spatiotemporal epidemic
    spread using stochastic models. Phytopathology 87, 139-46.
    \href{http://dx.doi.org/10.1094/PHYTO.1997.87.2.139}{doi:10.1094/PHYTO.1997.87.2.139}
}
\usage{
citrus_ctv
}
\description{
CTV incidence data for three orchards in eastern Spain reported for
consecutive years.
}
\details{
Both \code{IVI3and4} and \code{IVI6and7} orchards contained 216 trees, and
\code{El_Realengo} orchard, 400 trees.

From left to right, mappings represent infections dis-
covered up to May 1991 (45 infections), May 1992 (77 infections), and May 1993 (127 infections), respectively.

From left to right, mappings represent infections dis-
covered up to May 1991 (45 infections), May 1992 (77 infections), and May 1993 (127 infections), respectively. B, Corresponding information for epidemic
IVIA6&7 (13). From left to right, mappings represent infections discovered up to May 1990 (60 infections), May 1991 (77 infections), and May 1992 (127
infections), respectively.

Plots IVIA3&4 and IVIA6&7, consisted of 216 trees each of Washington navel orange on Troyer citrange planted in 1978
on a 2 x 6-m spacing. Both plots were located on the grounds of the Instituto Valenciano de Investigaciones
Agrarias, near Moncada, Valencia, Spain and were assayed yearly from 1981 to 1994.
}
\keyword{datasets}
