% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{clump}
\alias{clump}
\alias{clump.intensity}
\title{Regroup observational data into even clumps of individuals}
\usage{
clump(object, ...)

\method{clump}{intensity}(object, unit_size, fun = sum, ...)
}
\arguments{
\item{object}{An \code{intensity} object.}

\item{...}{Optional arguments to \code{fun}.}

\item{unit_size}{Size of a group unit. It must be a named vector, with names
corresponding to non-observational variables (i.e. space and time
variables). If the size of a variable in the data set is not a multiple
of the provided value in \code{unit_size}, some sampling units (the last
ones) will be dropped so that clumps of individuals remain even
throughout the data set.}

\item{fun}{Function used to group observational data together.}

\item{group_by}{Not yet implemented.}
}
\description{
This function provides a easy way to regroup recorded data into groups of
same number of individuals.
}
\examples{
my_data1 <- incidence(tomato_tswv$field_1929)
summary(my_data1)
plot(my_data1)

my_data2 <- clump(my_data1, unit_size = c(x = 3, y = 3))
summary(my_data2)
plot(my_data2)

my_data3 <- clump(my_data1, unit_size = c(t = 3), fun = mean)
summary(my_data3)
plot(my_data3)

}
