% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codling_moths}
\alias{codling_moths}
\title{Codling moth counts}
\format{A data frame with 30 rows and 3 variables:
\tabular{lll}{
    [, 1:2] \tab x,y \tab Spatial coordonates. \cr
    [, 3]   \tab d   \tab Counts. \cr
}}
\source{
Lavigne C, Ricci B, Franck P, Senoussi R. 2010. Spatial analyses of
    ecological count data: A density map comparison approach. Basic and
    Applied Ecology 11: 734-42.
    \href{http://dx.doi.org/10.1016/j.baae.2010.08.011}{10.1016/j.baae.2010.08.011}
}
\usage{
codling_moths
}
\description{
TODO.
}
\keyword{datasets}
