% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-factory.R
\name{coef.smle}
\alias{coef.smle}
\title{Extract Model Coefficients}
\usage{
\method{coef}{smle}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\value{
Coefficients extracted from the model object \code{object}.

  For standard model fitting classes this will be a named numeric vector.
  For \code{"maov"} objects (produced by \code{\link{aov}}) it will be a matrix.
}
\description{
\code{coef} is a generic function which extracts model coefficients
  from objects returned by modeling functions.  \code{coefficients} is
  an \emph{alias} for it.
}
\keyword{internal}
