% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dogwood_anthracnose}
\alias{dogwood_anthracnose}
\title{Dogwood anthracnose data.}
\format{An object of class \code{data.frame} with 329 rows and 3 columns.}
\source{
Zarnoch SJ, Anderson RL, Sheffield RM. 1995. Using the β-binomial
    distribution to characterize forest health. Canadian journal of forest
    research 25, 462–469.
}
\usage{
dogwood_anthracnose
}
\description{
Dogwood trees with symptoms of dogwood anthracnose. The assessments of disease
incidence were based on plots containing 10 dogwood trees and reported by Zarnoch et al (1995).
}
\keyword{datasets}
