% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiphy.R
\docType{package}
\name{epiphy}
\alias{epiphy}
\alias{epiphy-package}
\title{\code{epiphy}: Explore and analyze observations of plant disease epidemics}
\description{
This package is intended primarily for the plant pathologist community.
}
