% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{fisher}
\alias{fisher}
\alias{fisher.default}
\alias{fisher.count}
\alias{fisher.incidence}
\title{Fisher's index of dispersion, D}
\usage{
fisher(x, ...)

\method{fisher}{default}(x, flavor = c("count", "incidence"), n = NULL, ...)

\method{fisher}{count}(x, ...)

\method{fisher}{incidence}(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Not yet implemented.}

\item{flavor}{Which flavor of this index must be calculated ("count" or
"incidence")?}

\item{n}{Number of individuals per sampling unit. If \code{n} is provided,
the "incidence" flavor is calculated whatever the value of \code{flavor}.
Note that current implementation only deals with equal size sampling
units.}
}
\description{
In case of a count, this index corresponds to the ratio of the observed
variance to the observed mean. For a binary variable, a similar index can be
calculated using instead the ratio of the observed variance to the
theoretical variance if data follow a binomial law (i.e. a reference
distribution for a random pattern of individuals within sampling units).
}
\details{
Values of such an index can be interpreted as follows:
\itemize{
    \item D < 1: uniform or regular pattern;
    \item D = 1: random pattern;
    \item D > 1: aggregated or clustered pattern.
}
}
\examples{
# Count flavor:
## Using a standard R idiom:
fisher(aphids$i)

## Using a more "epiphy-like" approach:
fisher(count(aphids))

# Incidence flavor:
## Using a standard R idiom:
fisher(tobacco_viruses$i, n = tobacco_viruses$n)

## Using a more "epiphy-like" approach:
fisher(incidence(tobacco_viruses))

}
\references{
Fisher RA. 1925. Statistical methods for research workers. Oliver and Boyd,
    Edinburgh.
}
\seealso{
\code{\link{lloyd}}, \code{\link{morisita}}
}
