% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr-fitting.R
\name{fitDistr}
\alias{fitDistr}
\alias{fitDistr.incidence}
\title{Try to fit distributions to data sets}
\usage{
fitDistr(object, ...)

\method{fitDistr}{incidence}(object, ..., progress = TRUE, simulatePValue)
}
\arguments{
\item{x}{A data of the class \code{Intensity}.}

\item{type}{The kind of distribution to fit: "random", "aggregated" or "both".}
}
\value{
An object of class \code{DistFitting}, with the following components:

\tabular{ll}{
    \code{from} \tab Kind of the \code{Intensity} object. \cr
    \code{random} \tab Theoretical data under random assumption. \cr
    \code{aggregated} \tab Theoretical data under aggregated assumption. \cr
    \code{freqTable} \tab Entry 4 \cr
}
}
\description{
Depending of the nature of the data submitted, different distributions are
tryied to be fitted to the data.
}
\details{
Two distributions are used. One under the assumption of no aggregation and
the other one under the assumption of aggregation.
}
\section{Methods (by class)}{
\itemize{
\item \code{incidence}: For incidence data, the two distributions fitted are binomial and beta-binomial
distributions.
}}

\examples{
\dontrun{
data <- Incidence(Cochran1936)
fitDistr(data)
}

}
\references{

}
