% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hop_viruses}
\alias{hop_viruses}
\title{Incidence of three viruses in an Australian hop garden}
\format{Three data sets, one for each virus. The structure is the same for each virus:
List elements are "HpLV", "HpMV", and "ApMV".

A data frame with ... rows and ... variables:
\tabular{lll}{
    [, 1:2] \tab x,y   \tab Grid coordonates. \cr
    [, 3:4] \tab xm,ym \tab Metric coordinates. \cr
    [, 5]   \tab t     \tab Year of assays for viruses. \cr
    [, 6]   \tab r     \tab Incidence. 1 = infected, 0 = uninfected. \cr
    [, 7]   \tab n     \tab Sampling unit size. n = 1 means that the sampling
                            unit size is the plant. \cr
}}
\source{
Pethybridge SJ, Madden LV. 2003. Analysis of spatiotemporal dynamics
    of virus spread in an Australian hop garden by stochastic modeling. Plant
    Disease 87:56-62.
    \href{http://dx.doi.org/10.1094/PDIS.2003.87.1.56}{doi:10.1094/PDIS.2003.87.1.56}
}
\usage{
hop_viruses
}
\description{
TODO. Three viruses: Hop latent virus (HpLV), Hop mosaic virus (HpMV), and Apple mosaic virus (ApMV).
Hop (Humulus lupulus) cv. Victoria, in a commercial hop farm at Bushy Park, Tasmania, Australia.
Established in 1989. An Australian hop garden.
51 plants within each roaw and 25 rows (total number of plants = 1,275 plants).
1.8 m within rows, and 2.1 m between rows.
}
\keyword{datasets}
