% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\alias{fisher1}
\alias{morisita1}
\alias{iwao}
\title{Aggregation indices}
\usage{
fisher1(data, confLevel = 0.95)

morisita1(data)

iwao(data)
}
\arguments{
\item{data}{A vector of integers or an \code{Intensity} object.}
}
\description{
Compute the eponym indices of aggregation.
}
\section{Functions}{
\itemize{
\item \code{fisher1}: Fisher's index of dispersion, D

In case of a binary variable, the index is a ratio of two variances: the
observed variance and the theoretical variance if data follow a binomial law,
i.e. a reference distribution for a random pattern of diseased individuals
within sampling units.
Under the null hypothesis (D = 1, random pattern), (N - 1)D follows a
chi-squared distribution with N - 1 degrees of freedom. N is the number of
quadrats.

Count (Moradi-Vajargah et al, 2011: CAPS)... cf Gosme for Incidence
V/M < 1: means uniform / REGULAR
V/M = 1: means RANDOM
V/M > 1: means AGGREGATED / clustered

\item \code{morisita1}: Morisita's coefficient of dispersion

Interpretation for incidence data remains uncertain, except for low p (Madden and Hughes, 1995)

\item \code{iwao}: Iwao's patchiness regression index
}}

\references{
Fisher RA. 1925. Statistical methods for research workers. Oliver and Boyd, Edinburgh.

Morisita M. 1962. Iδ-Index, a measure of dispersion of
individuals. Researches on Population Ecology 4, 1–7.
\href{http://dx.doi.org/doi:10.1007/BF02533903}{doi:10.1007/BF02533903}

Madden LV, Hughes G. 1995. Plant disease incidence: Distributions,
heterogeneity, and temporal analysis. Annual Review of Phytopathology 33 (1):
529–64.
\href{http://dx.doi.org/doi:10.1146/annurev.py.33.090195.002525}{doi:10.1146/annurev.py.33.090195.002525}

Iwao S. 1968. A new regression method for analyzing the aggregation pattern
of animal populations. Researches on Population Ecology 10, 1–20.
\href{http://dx.doi.org/10.1007/BF02514729}{doi:10.1007/BF02514729}
}
