% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{is.intensity}
\alias{is.intensity}
\alias{is.count}
\alias{is.incidence}
\alias{is.severity}
\title{Test if an object is of class \code{intensity} or one of its subclasses.}
\usage{
is.intensity(x)

is.count(x)

is.incidence(x)

is.severity(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if its argument's value is the corresponding \code{intensity}
    object and FALSE otherwise.
}
\description{
Test if an object is of class \code{intensity} or one of its subclasses
(i.e. \code{count}, \code{incidence} or \code{severity}).
}
