% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{iwao}
\alias{iwao}
\title{indices Iwao's patchiness regression index}
\usage{
iwao(data)
}
\arguments{
\item{data}{Data.}
}
\description{
TODO: Not yet implemented.
}
\references{
Iwao S. 1968. A new regression method for analyzing the aggregation pattern
of animal populations. Researches on Population Ecology 10, 1–20.
\href{http://dx.doi.org/10.1007/BF02514729}{doi:10.1007/BF02514729}
}
