% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link}
\alias{link}
\alias{logit}
\alias{probit}
\alias{cloglog}
\title{Some link functions.}
\usage{
logit(x, rev = FALSE)

probit(x, rev = FALSE)

cloglog(x, rev = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{rev}{The inverse of the function?}
}
\value{
A numeric vector.
}
\description{
Logit, probit and cloglog functions are available.
The logit and the logistic (with rev = TRUE), i.e. the inverse-logit functions.
Probit is a wrapper around \code{qnorm} (for \eqn{probit}) and \code{pnorm} (for \eqn{probit^{-1}})
Complementary log-log transformation.
}
