% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{lloyd}
\alias{lloyd}
\alias{lloyd.default}
\alias{lloyd.count}
\title{indices Lloyd's index of patchiness or Lloyd's mean crowding}
\usage{
lloyd(x, type)

\method{lloyd}{default}(x, type = c("patchiness", "mean-crowding"))

\method{lloyd}{count}(x, type = c("patchiness", "mean-crowding"))
}
\description{
Returns a value between 0 and \code{n}, the size of a sampling unit. A
value lower than 1 indicates a random pattern, while a value greater than 1
indicates an aggregated pattern. The degree of aggregation increases with
the value of the index.
}
\details{
\itemize{
    \item LIP < 1: means REGULAR
    \item LIP = 1: means RANDOM
    \item LIP > 1: means AGGREGATED
}
}
\references{
Lloyd M. 1967. Mean crowding. The Journal of Animal Ecology 36,
    1–30.
}
\seealso{
\code{\link{fisher}}, \code{\link{morisita}}
}
