% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-factory.R
\name{logLik.smle}
\alias{logLik.smle}
\title{Extract log-likelihood}
\usage{
\method{logLik}{smle}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\value{
Returns an object of class \code{logLik}.  This is a number with at
  least one attribute, \code{"df"} (\bold{d}egrees of \bold{f}reedom),
  giving the number of (estimated) parameters in the model.

  There is a simple \code{print} method for \code{"logLik"} objects.

  There may be other attributes depending on the method used: see the
  appropriate documentation.  One that is used by several methods is
  \code{"nobs"}, the number of observations used in estimation (after
  the restrictions if \code{REML = TRUE}).
}
\description{
This function returns the maximal log-likelihood estimated with
\code{\link{smle}}, if \code{f} returned log-likelihood value.
}
\keyword{internal}
