% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomp.R
\name{mapcomp}
\alias{mapcomp}
\alias{mapcomp.data.frame}
\alias{mapcomp.matrix}
\alias{mapcomp.count}
\alias{mapcomp.incidence}
\title{Map Comparison procedure.}
\usage{
mapcomp(data, ...)

\method{mapcomp}{data.frame}(data, delta, h, nperm = 100,
  edge_correction = TRUE, threads = 1)

\method{mapcomp}{matrix}(data, delta, h, nperm = 100,
  edge_correction = TRUE, threads = 1)

\method{mapcomp}{count}(data, delta, h, nperm = 100, edge_correction = TRUE,
  threads = 1)

\method{mapcomp}{incidence}(data, delta, h, nperm = 100,
  edge_correction = TRUE, threads = 1)
}
\arguments{
\item{data}{A data frame with 3 columns.}
}
\description{
\code{mapcomp} procedure. Spatial analyses of ecological count data: A density map comparaison approach
}
\details{
the Hellinger distance.
}
\examples{
my_res <- mapcomp(codling_moths, 4, 9, edge_correction = FALSE)
plot(my_res)

}
\references{
Lavigne C, Ricci B, Franck P, Senoussi R. 2010. Spatial analyses of
ecological count data: A density map comparison approach. Basic and Applied
Ecology. 11:734–742.
}
