% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{mapv}
\alias{mapv}
\alias{mapv<-}
\title{Existing variable mappings}
\usage{
mapv(x)

mapv(x, keep = TRUE) <- value
}
\arguments{
\item{x}{An \code{intensity} object.}

\item{value}{A \code{mapping} object.}

\item{keep}{Logical. Do we keep any previous mapped variables that are not
redifined in the \code{mapping} object?}
}
\description{
Get or set existing variable mappings.
}
\examples{
my_data <- count(aphids)
my_data
#... TODO
# x, y, X, Y
mapv(my_data)
mapv(my_data) <- mapping(x = X, y = Y)
mapv(my_data)
mapv(my_data) <- mapping(x = x, r = r, keep = FALSE)
mapv(my_data)

}
\seealso{
\code{\link{mapping}}
}
