% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{morisita}
\alias{morisita}
\alias{morisita.default}
\alias{morisita.count}
\alias{morisita.incidence}
\title{indices Morisita's coefficient of dispersion}
\usage{
morisita(x, ...)

\method{morisita}{default}(x, flavor = c("count", "incidence"), n = NULL,
  ...)

\method{morisita}{count}(x, ...)

\method{morisita}{incidence}(x, ...)
}
\description{
Interpretation for incidence data remains uncertain, except for low p (Madden and Hughes, 1995)
}
\references{
Morisita M. 1962. Iδ-Index, a measure of dispersion of
individuals. Researches on Population Ecology 4, 1–7.
\href{http://dx.doi.org/doi:10.1007/BF02533903}{doi:10.1007/BF02533903}

Madden LV, Hughes G. 1995. Plant disease incidence: Distributions,
heterogeneity, and temporal analysis. Annual Review of Phytopathology 33 (1):
529–64.
\href{http://dx.doi.org/doi:10.1146/annurev.py.33.090195.002525}{doi:10.1146/annurev.py.33.090195.002525}
}
\seealso{
\code{\link{fisher}}, \code{\link{lloyd}}
}
