% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{offspring_survival}
\alias{offspring_survival}
\title{Offspring survival of rats experiencing different diets.}
\format{
A data frame with 32 rows and 3 variables:
\tabular{rll}{
    [, 1]   \tab group \tab Either control or treated group. \cr
    [, 2]   \tab i     \tab Pups weaned. \cr
    [, 3]   \tab n     \tab Pups alive at 4 days. \cr
}
}
\source{
Weil CS. 1970. Selection of the valid number of sampling units and a
    consideration of their combination in toxicological studies involving
    reproduction, teratogenesis or carcinogenesis. Food and Cosmetics
    Toxicology 8: 177-182.
}
\usage{
offspring_survival
}
\description{
Results of an experiment where two groups of 16 female rats were fed
different diets during pregnancy and lactation periods. One group's diet
contained a chemical under review, and the other one was a control. For each
litter, the number of pups alive at 4 days, and the number of pups weaned
(i.e. that survived the 21-day lactation period) were recorded.
}
\keyword{datasets}
