% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onion_bacterial_blight}
\alias{onion_bacterial_blight}
\title{Bacterial blight of onion.}
\format{A data frame with 1134 rows and 5 variables (x, y, t, d, n) d- record
range(x) = 1:27
range(y) = 1:21}
\source{
Roumagnac P, Pruvost O, Chiroleu F, Hughes G. 2004. Spatial and
    temporal analyses of bacterial blight of onion caused by Xanthomonas
    axonopodis pv. allii. Phytopathology 94, 138–146.
    \href{http://dx.doi.org/10.1094/PHYTO.2004.94.2.138}{doi:10.1094/PHYTO.2004.94.2.138}
}
\usage{
onion_bacterial_blight
}
\description{
Map of bacterial blight of onion incidence at two assessment dates.
21 x 21 grid???
}
\details{
Experimental plots were sown with the same naturally X. axonopodis pv. allii-contaminated onion
(A. cepa L. cv. Chateau-vieux) seed lot. The contamination rate of seed was about 0.04%.
}
\keyword{datasets}
