% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{plot.power_law}
\alias{plot.power_law}
\title{Plot results of a power law analysis}
\usage{
\method{plot}{power_law}(x, ..., scale = c("logarithmic", "linear"),
  observed = TRUE, model = TRUE, random = TRUE)
}
\arguments{
\item{x}{A \code{\link{power_law}} object.}

\item{...}{Additional arguments to be passed to other methods.
TODO: Dashed lines indicate the cases where both variances are equal, which suggests an absence of aggregation.
Points should lie on this line if ...}

\item{scale}{Logarithmic or standard linear scale to display the results?}

\item{observed}{Logical.}

\item{model}{Logical.}

\item{random}{Logical. Theoretical Random distribution.}
}
\description{
Plot results of a power law analysis.
}
