% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{powerLaw}
\alias{powerLaw}
\title{Taylor's and Binomial Power Laws}
\usage{
powerLaw(x, baseLog = 10, ...)
}
\description{
These functions allow to...
}
\details{
\deqn{log_{10}(s_y^{2}) = log_{10}(a) + b log_{10}(Y)}
with \eqn{Y} is equivalent to \eqn{\bar{y}(1 - \bar{y})} in case of incidence
data, and \eqn{\bar{y}} in case of count data. \eqn{y} correspond to the
proportion of diseased plants or plant parts in each sampling unit, whereas
\eqn{y} corresponds to the absolute value in case of data count.
}
