% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyrethrum_ray_blight}
\alias{pyrethrum_ray_blight}
\title{Ray blight disease of pyrethrum.}
\format{A data frame with 62 rows and 2 variables:
\tabular{rll}{
    [, 1] \tab r \tab Number of diseased samples in the quadrat (i.e., 0, 1, 2, 3, 4, 5, or 6). \cr
    [, 2] \tab n \tab Sampling unit size. Number of samples assessed per quadrat, n = 6. \cr
}}
\source{
Pethybridge SJ, Esker P, Hay F, Wilson C, Nutter FW. 2005.
    Spatiotemporal description of epidemics caused by Phoma ligulicola in
    Tasmanian pyrethrum fields. \emph{Phytopathology} 95, 648–658.
    \href{http://dx.doi.org/10.1094/PHYTO-95-0648}{doi:10.1094/PHYTO-95-0648}
}
\usage{
pyrethrum_ray_blight
}
\description{
An assessment of the incidence of foliar symptoms due to ray blight disease
of pyrethrum in 62 quadrats.
}
\keyword{datasets}
