% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{regroup}
\alias{regroup}
\alias{regroup.intensity}
\title{regroup data}
\usage{
regroup(object, ...)

\method{regroup}{intensity}(object, unitDim, groupBy, fun = sum,
  keepLabels = TRUE)
}
\arguments{
\item{unitDim}{Dimensions of the base unit. All the dimensions must be an
integer diviser of the size of the object. Basically should be a list of two named
element: space and time, each one speciffying the dimensions. If not a list,
must be the same size as \code{length(object@space, object@time)}.}

\item{groupBy}{Parameter in the form of c("", "", ...). A vector of names}

\item{fun}{The function to use.}
}
\description{
Useful.
}
\examples{
myData <- incidence(Cochran1936)
myData <- regroup(myData, unitDim = c(3, 3, 1))
myData <- as.severity(myData)

## The folowing code do not work because it leads to non-integer number
## which is imcomatible with the definition of incidence
\dontrun{
myData <- regroup(myData, unitDim = c(3, 3, 1), fun = mean)
}
## To perform the desired calculation, we need to do:
## Because we go out of the strict definition of the class incidence
myData <- incidence(Cochran1936)
myData <- regroup(myData, unitDim = c(3, 3, 1))
## With dplyr library:
as.data.frame(myData) \%>\% dplyr::summarise(mean = r / n)
## Or with a vanilla synthax:
myData <- as.data.frame(myData)
myData$mean <- myData$r / myData$n
}
