% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-dispersal-model.R
\docType{data}
\name{simple_model}
\alias{simple_model}
\title{Simple dispersal model}
\format{An object of class \code{list} of length 6.}
\usage{
simple_model
}
\arguments{
\item{nfoci}{The number of (initial) infections.}

\item{xrate}{Multiplication rate, the number of new infections produced by
each pre-existing infection at each generation.}

\item{ngen}{The number of generations.}

\item{lambda}{The rate parameter for the exponential distribution.}
}
\description{
This model is used to perform some assessments and test statistical methods
implemented in the \code{epiphy} package.
}
\examples{
# To use the "hiden" functions of this simple model:
invisible(list2env(epiphy:::simple_model, environment()))

set.seed(12345)
foci <- disperse(nfoci = 10, xrate = 20, ngen = 3, lambda = 50, ngen_active = 1)
plot(foci, cex = 0.1)
quad <- quadrat(surf_dim = c(x = 1, y = 1), nint = c(x = 90, y = 90))
collection <- collect(foci, quad)

# To perform analyses:
collection$n <- 1
my_data <- incidence(collection, mapping(r = incidence))
my_data <- clump(my_data, unit_size = c(x = 3, y = 3)) # /!\\ avec clump(...), on retire les colonnes non mappées
res <- fit_distr(my_data)
res
summary(res)
plot(res)

lambdas <- c(5, 10, 20, 40, 80)

# Multi-data set analyses
set.seed(12345)
quad <- quadrat(surf_dim = c(x = 1, y = 1), nint = c(x = 90, y = 90))
my_data <- list()
for (i in 1:100) {
    nfoci <- sample(1:100, 1)
    xrate <- sample(5:15, 1)
    print(paste0("set ", i, " to be done (nfoci = ", nfoci, "; xrate = ", xrate,")"))
    foci <- disperse(nfoci = nfoci, xrate = xrate, ngen = 2, lambda = 20)
    my_data[[i]] <- collect(foci, quad)
}
# saveRDS(my_data, "simple_model_data_final_5-15.rds")

## Power law
my_data2 <- lapply(my_data, function(x) {
    x$n <- 1
    x <- incidence(x, mapping(r = incidence))
    clump(x, unit_size = c(x = 3, y = 3))
})

res <- power_law(my_data2)
res
summary(res)
plot(res)

## Spatial hierarchy
low <- my_data2
high <- lapply(my_data2, level_up)

res <- spatial_hier(low, high)
res
summary(res)
plot(res)

}
\references{
Esker PD, Sparks AH, Antony G, Bates M, Dall' Acqua W, Frank EE, Huebel L,
Segovia V, Garrett KA, 2007. Ecology and Epidemiology in R: Modeling
dispersal gradients. The Plant Health Instructor.
\href{http://dx.doi.org/10.1094/PHI-A-2007-1226-03}{doi:10.1094/PHI-A-2007-1226-03}
}
\keyword{internal}
