% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_epidemics}
\alias{simulated_epidemics}
\title{Examples of simulated epidemic data.}
\format{A data frame with 864 rows and 6 variables:
\tabular{rll}{
    [, 1]   \tab pattern \tab Either clumped (i.e. aggregated), random or regular. \cr
    [, 2]   \tab mu \tab Median spore dispersal parameter. \cr
    [, 3:4] \tab x,y \tab Spatial coordinates. \cr
    [, 5]   \tab d   \tab Diseased individuals (from 0 to 100). \cr
    [, 6]   \tab n   \tab Sampling unit size. Here, n = 100 individuals per sampling unit. \cr
}}
\source{
Xu XM, Madden LV. 2004. Use of SADIE statistics to study spatial
  dynamics of plant disease epidemics. Plant Pathology 53, 38–49.
  \href{http://dx.doi.org/10.1111/j.1365-3059.2004.00949.x}{doi:10.1111/j.1365-3059.2004.00949.x}
}
\usage{
simulated_epidemics
}
\description{
Epidemics generated using the stochastic simulator of Xu and Madden (2004).
The data are number of diseased plants per quadrat (out of a total of n = 100
plant in each). N = 144 quadrats. Differents parameters (pattern and mu).
}
\keyword{datasets}
