% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-hier.R
\name{spatial_hier}
\alias{spatial_hier}
\title{Spatial hierarchy analysis}
\usage{
spatial_hier(low, high)
}
\arguments{
\item{low}{An list of \code{intensity} objects.}

\item{high}{An list of \code{intensity} objects.}
}
\description{
TODO
}
\examples{
my_data_low  <- incidence(tomato_tswv$field_1928)
# TODO: 2 bugs to correct here (before)
# my_data_low  <- split(my_data_low, by = "t")[[1]]
# my_data_high <- clump(my_data_low, unit_size = c(x = 3, y = 3))

my_data_low <- incidence(tomato_tswv$field_1929)
my_data_low <- clump(my_data_low, c(x = 3, y = 3))
my_data_high <- my_data_low
my_data_high$data$n <- 1
my_data_high$data$r <- ifelse(my_data_high$data$r > 0, 1, 0)
my_data_low  <- split(my_data_low, by = "t")
my_data_high <- split(my_data_high, by = "t")
res <- spatial_hier(my_data_low, my_data_high)

res
summary(res)
plot(res)

}
