% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{split}
\alias{split}
\alias{split.incidence}
\title{Divide into groups}
\usage{
\method{split}{incidence}(x, f, drop, ...)
}
\arguments{
\item{x}{intensity object containing values to be divided into groups.}

\item{f}{A \code{factor} in the sense that \code{\link{as.factor}(f)} defines the grouping.}
}
\description{
Great function
}
\examples{
myData <- incidence(Cochran1936)

res1 <- split(myData, myData@time[, 1])
# is equivalent to:
res2 <- lapply(levels(as.factor(myData@time[, 1])),
               function(i) myData[t = i])
names(res2) <- levels(as.factor(myData@time[, 1]))

}
