% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_viruses}
\alias{tobacco_viruses}
\title{Incidence of tobacco plants infected with viruses}
\format{A data frame with 75 rows and 2 variables:
\tabular{rll}{
    [, 1] \tab r \tab Disease incidence. 1: Diseased, 0: Healthy. \cr
    [, 2] \tab n \tab Sampling unit size. n = 1 means that the sampling unit size is the plant. \cr
}}
\source{
Madden LV, Pirone TP, Raccah B. 1987. Analysis of spatial patterns of
    virus-diseased tobacco plants. Phytopathology 77, 1409–1417.
}
\usage{
tobacco_viruses
}
\description{
Field with N = 75 sampling units with n = 40 plants in each one.
}
\keyword{datasets}
