% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{a2a}
\alias{a2a}
\title{a2a: A wAy to pAinlessly switch between different power LAw formulAtions}
\usage{
a2a(intercept, from = c("Ar", "ar", "AR", "aR"), to = c("Ar", "ar", "AR",
  "aR"), slope, n)
}
\arguments{
\item{intercept}{Intercept parameter to be converted.}

\item{from}{Kind of the input intercept parameter.}

\item{to}{Desired kind for the ouput intercept parameter.}

\item{slope}{Slope parameter.}

\item{n}{Number of individuals per sampling unit.}
}
\description{
\code{a2a} was designed to avoid headaches that are likely to occur when
working with different formulations of the binomial power law analysis.
}
\details{
The binomial power law can be expressed as: \eqn{s_y^2 = (intercept)(s_{bin}^2)^b}.
But different forms of (intercept) are possible depending on the formulation of the
binomial power law.
\tabular{ccccc}{
      \tab Ar         \tab ar      \tab AR         \tab aR      \cr
   Ar \tab 1          \tab n^b     \tab n^{2(b-1)} \tab n^{b-2} \cr
   ar \tab n^{-b}     \tab 1       \tab n^{b-2}    \tab n^{-2}  \cr
   AR \tab n^{2(1-b)} \tab n^{2-b} \tab 1          \tab n^{-b}  \cr
   aR \tab n^{2-b}    \tab n^2     \tab n^b        \tab 1       \cr
}
}
\examples{

a2a(from = , to = , n = , b = )
a2a(to = , data = <powerLaw object>)

}
