% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{a2a}
\alias{a2a}
\alias{a2a.numeric}
\alias{a2a.list}
\title{a2a: A wAy to pAinlessly switch between different power LAw formulAtions}
\usage{
a2a(x, ...)

\method{a2a}{numeric}(x, slope, n, from = c("Ar", "ar", "AR", "aR"),
  to = c("Ar", "ar", "AR", "aR"))

\method{a2a}{list}(x, to = c("Ar", "ar", "AR", "aR"))
}
\arguments{
\item{x}{Intercept parameter to be converted or a named list with the
parameter to be converted ("Ar", "ar", "AR" or "aR"), the slope
("slope"), and the number of individual per sampling unit ("n").}

\item{slope}{Slope parameter.}

\item{n}{Number of individuals per sampling unit.}

\item{from}{Kind of the input intercept parameter ("Ar", "ar", "AR" or "aR").}

\item{to}{Desired kind for the ouput intercept parameter ("Ar", "ar", "AR" or
"aR").}
}
\description{
\code{a2a} was designed to avoid headaches that are likely to occur when
working with different formulations of the binomial power law analysis.
}
\details{
The binomial power law can be expressed as: \eqn{s_y^2 = (intercept)(s_{bin}^2)^b}.
But different forms of (intercept) are possible depending on the formulation of the
binomial power law.
\tabular{ccccc}{
      \tab Ar         \tab ar      \tab AR         \tab aR      \cr
   Ar \tab 1          \tab n^b     \tab n^{2(b-1)} \tab n^{b-2} \cr
   ar \tab n^{-b}     \tab 1       \tab n^{b-2}    \tab n^{-2}  \cr
   AR \tab n^{2(1-b)} \tab n^{2-b} \tab 1          \tab n^{-b}  \cr
   aR \tab n^{2-b}    \tab n^2     \tab n^b        \tab 1       \cr
}
}
\examples{
# Values from the power_law() example:
Ar    <- 38.6245
slope <- 1.9356
n     <- 9

# Usual function call syntax:
a2a(Ar, slope, n, from = "Ar", to = "ar")

# Other syntaxes:
inputs <- list(Ar = Ar, slope = slope, n = n)
a2a(inputs, "ar")
require(magrittr)
inputs \%>\% a2a("ar")

}
