% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aphids}
\alias{aphids}
\title{Counts of aphids.}
\format{A data frame with 63 rows and 3 variables:
\tabular{lll}{
    [, 1:2] \tab x,y   \tab Grid spatial coordinates. \cr
    [, 3:4] \tab xm,ym \tab Metric spatial coordonates. \cr
    [, 5]   \tab i     \tab Counts of aphids. \cr
}}
\source{
Perry JN, Winder L, Holland JM, Alston RD. 1999. Red-blue plots for
    detecting clusters in count data. Ecology Letters 2, 106-13.
    \href{http://dx.doi.org/10.1046/j.1461-0248.1999.22057.x}{doi:10.1046/j.1461-0248.1999.22057.x}
}
\usage{
aphids
}
\description{
Counts of 554 aphids of the species Sitobion avenae, sampled on 28 June 1996
in a 250 x 180-m field of winter wheat near Wimborne, Dorset, UK. The 63
sampling units, made of the inspection of five tillers each, were located on
a 9 x 7 rectangular grid at intervals of 30 m.
}
\keyword{datasets}
