% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citrus_ctv}
\alias{citrus_ctv}
\title{CTV disease incidence at 3 assessment dates.}
\format{There 3 data sets here:

Field IVIA3&4: A data frame with ... rows and 7 variables:
\tabular{rll}{
    [, 1:2] \tab x,y \tab Spatial coordinates. \cr
    [, 3:4] \tab X,Y \tab Metric spatial coordinates. \cr
    [, 5]   \tab t   \tab Time/Year of the disease assessments (1991, 1992, 1993 or 1994). \cr
    [, 6]   \tab r   \tab Disease incidence records. 1: Diseased, 0: Healthy. \cr
    [, 7]   \tab n   \tab Sampling unit size. n = 1 means that the sampling unit size is the plant. \cr
}

Field IVIA6&7: A data frame with ... rows and 7 variables:
\tabular{rll}{
    [, 1:2] \tab x,y \tab Spatial coordinates. \cr
    [, 3:4] \tab X,Y \tab Metric spatial coordinates. \cr
    [, 5]   \tab t   \tab Time/Year of the disease assessments (1990, 1991, or 1992). \cr
    [, 6]   \tab r   \tab Disease incidence records. 1: Diseased, 0: Healthy. \cr
    [, 7]   \tab n   \tab Sampling unit size. n = 1 means that the sampling unit size is the plant. \cr
}

Field El-Realengo: A data frame with ... rows and 7 variables:
\tabular{rll}{
    [, 1:2] \tab x,y \tab Spatial coordinates. \cr
    [, 3:4] \tab X,Y \tab Metric spatial coordinates. \cr
    [, 5]   \tab t   \tab Time/Year of the disease assessments (1981, 1982, 1984, 1985, 1990). \cr
    [, 6]   \tab r   \tab Disease incidence records. 1: Diseased, 0: Healthy. \cr
    [, 7]   \tab n   \tab Sampling unit size. n = 1 means that the sampling unit size is the plant. \cr
}}
\source{
Gottwald TR, Cambra M, Moreno P, Camarasa E, Piquer J. 1996. Spatial
    and temporal analyses of citrus tristeza virus in eastern Spain.
    Phytopathology 86, 45–55.

Gibson GJ. 1997. Investigating mechanisms of spatiotemporal epidemic
    spread using stochastic models. Phytopathology 87, 139-46.
    \href{http://dx.doi.org/10.1094/PHYTO.1997.87.2.139}{doi:10.1094/PHYTO.1997.87.2.139}
}
\usage{
citrus_ctv
}
\description{
Map of Citrus tristeza virus disease incidence at three assessment years.
21 x 21 grid???
}
\details{
Epidemic of citrus tristeza virus (CTV) observed in eastern Spain (epidemic IVIA3&4 [13]). Black squares denote positions of infections observed at
previous observation times, and white squares represent the new infections recorded in each disease map. From left to right, mappings represent infections dis-
covered up to May 1991 (45 infections), May 1992 (77 infections), and May 1993 (127 infections), respectively.

A, Epidemic of citrus tristeza virus (CTV) observed in eastern Spain (epidemic IVIA3&4 [13]). Black squares denote positions of infections observed at
previous observation times, and white squares represent the new infections recorded in each disease map. From left to right, mappings represent infections dis-
covered up to May 1991 (45 infections), May 1992 (77 infections), and May 1993 (127 infections), respectively. B, Corresponding information for epidemic
IVIA6&7 (13). From left to right, mappings represent infections discovered up to May 1990 (60 infections), May 1991 (77 infections), and May 1992 (127
infections), respectively.

Plots IVIA3&4 and IVIA6&7, consisted of 216 trees each of Washington navel orange on Troyer citrange planted in 1978
on a 2 x 6-m spacing. Both plots were located on the grounds of the Instituto Valenciano de Investigaciones
Agrarias, near Moncada, Valencia, Spain and were assayed yearly from 1981 to 1994.
}
\keyword{datasets}
