% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codling_moths}
\alias{codling_moths}
\title{Count of codling moth larvae.}
\format{
A data frame with 30 rows and 3 variables:
\tabular{lll}{
    [, 1:2] \tab x,y \tab Metric spatial coordonates. \cr
    [, 3]   \tab i   \tab Counts of codling moth larvae. \cr
}
}
\source{
Lavigne C, Ricci B, Franck P, Senoussi R. 2010. Spatial analyses of
    ecological count data: A density map comparison approach. Basic and
    Applied Ecology 11: 734-42.
    \doi{10.1016/j.baae.2010.08.011}
}
\usage{
codling_moths
}
\description{
Codling moth diapausing larvae were collected in an apple orchard in
south-eastern France. Larvae were caught on strip traps wrapped around tree
trunks in July 2008 and collected the following October. 30 traps were used.
}
\keyword{datasets}
