% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dogwood_anthracnose}
\alias{dogwood_anthracnose}
\title{Incidence of dogwood anthracnose.}
\format{A data frame with 329 rows and 3 variables:
\tabular{rll}{
    [, 1] \tab t \tab Year of disease assessments (1990 or 1991).. \cr
    [, 2] \tab i \tab Number of diseased plants (from 0 to 10). \cr
    [, 3] \tab n \tab Sampling unit size. Here, n = 10 plants per sampling
                      unit (or plot). \cr
}}
\source{
Zarnoch SJ, Anderson RL, Sheffield RM. 1995. Using the \eqn{\beta}-binomial
    distribution to characterize forest health. Canadian journal of forest
    research 25, 462–469.
}
\usage{
dogwood_anthracnose
}
\description{
Incidence data from the Dogwood Anthracnose Impact Assessment Program for
1990 and 1991, in the Southeast of the USA, reported by Zarnoch et al (1995).
Only plots with exactly n = 10 dogwood trees are present in the data set (168
and 161 plots in 1990 and 1991, respectively).
}
\keyword{datasets}
