% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiphy.R
\docType{package}
\name{epiphy}
\alias{epiphy}
\alias{epiphy-package}
\title{\code{epiphy}: Explore and analyze observations of plant disease epidemics}
\description{
This package is intended primarily for the plant pathologist community.
}
\references{
Gosme M. 2008. Comment analyser la structure spatiale et modéliser le
développement spatio-temporel des épiphyties? Canadian Journal of Plant
Pathology, 30:4-23.

Madden LV, Hughes G, van den Bosch F. 2007. Spatial aspects of epidemics -
III: Patterns of plant disease. In: The study of plant disease epidemics,
235–278. American Phytopathological Society, St Paul, MN.
}
