% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{fisher}
\alias{fisher}
\alias{fisher.default}
\alias{fisher.count}
\alias{fisher.incidence}
\title{indices Fisher's index of dispersion, D}
\usage{
fisher(x, ...)

\method{fisher}{default}(x, flavor = c("count", "incidence"), n = NULL, ...)

\method{fisher}{count}(x, ...)

\method{fisher}{incidence}(x, ...)
}
\description{
In case of a binary variable, the index is a ratio of two variances: the
observed variance and the theoretical variance if data follow a binomial law,
i.e. a reference distribution for a random pattern of diseased individuals
within sampling units.
Under the null hypothesis (D = 1, random pattern), (N - 1)D follows a
chi-squared distribution with N - 1 degrees of freedom. N is the number of
quadrats.
}
\details{
Count (Moradi-Vajargah et al, 2011: CAPS)... cf Gosme for Incidence

\itemize{
    \item V/M < 1: means uniform / REGULAR
    \item V/M = 1: means RANDOM
    \item V/M > 1: means AGGREGATED / clustered
}
}
\examples{
# Base approach
library(dplyr)
my_data <- filter(tomato_tswv$field_1929) \%>\%
    filter(t == 1) \%>\%
    incidence() \%>\%
    clump(unit_dim = c(x = 3, y = 3))
r <- as.data.frame(my_data)[["i"]]
n <- as.data.frame(my_data)[["n"]][[1]]
fisher(r, n = n, flavor = "incidence")

# More elegant approach
my_data <- incidence(tomato_tswv$field_1929)
fisher(my_data)

}
\references{
Fisher RA. 1925. Statistical methods for research workers. Oliver and Boyd, Edinburgh.
}
\seealso{
\code{\link{lloyd}}, \code{\link{morisita}}
}
