% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyrethrum_ray_blight}
\alias{pyrethrum_ray_blight}
\title{Incidence of ray blight disease of pyrethrum.}
\format{
A data frame with 62 rows and 2 variables:
\tabular{rll}{
    [, 1] \tab i \tab Number of diseased plants (from 0 to 6). \cr
    [, 2] \tab n \tab Sampling unit size. Here, n = 6 plants per sampling
                      unit. \cr
}
}
\source{
Pethybridge SJ, Esker P, Hay F, Wilson C, Nutter FW. 2005.
    Spatiotemporal description of epidemics caused by Phoma ligulicola in
    Tasmanian pyrethrum fields. \emph{Phytopathology} 95, 648–658.
    \doi{10.1094/PHYTO-95-0648}
}
\usage{
pyrethrum_ray_blight
}
\description{
An assessment of the incidence of ray blight disease of pyrethrum in 62
sampling units, containing 6 plants each.
}
\keyword{datasets}
