% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{split.intensity}
\alias{split.intensity}
\title{Divide into groups and reassemble}
\usage{
\method{split}{intensity}(x, f, drop = FALSE, ..., by)
}
\arguments{
\item{x}{vector or data frame containing values to be divided into groups.}

\item{f}{a \sQuote{factor} in the sense that \code{\link{as.factor}(f)}
    defines the grouping, or a list of such factors in which case their
    interaction is used for the grouping.}

\item{drop}{logical indicating if levels that do not occur should be dropped
    (if \code{f} is a \code{factor} or a list).}

\item{...}{further potential arguments passed to methods.}

\item{by}{The name(s) of the variable(s) which define(s) the grouping.}
}
\description{
TODO
}
