% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-classes.R
\name{threshold}
\alias{threshold}
\title{Threshold : To go to higher level in the hierarchy}
\usage{
threshold(data, value = c(0, Inf))
}
\arguments{
\item{data}{An \code{intensity} object.}

\item{value}{All the intensity values lower or equal to this threshold
are set set to 0. The other values are set to 1.}
}
\description{
"The Threshold tool transforms the current layer or the selection into a
black and white image, where white pixels represent the pixels of the image
whose Value is in the threshold range (1), and black pixels represent pixels with
Value out of the threshold range (0)."
}
