% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-factory.R
\name{vcov.smle}
\alias{vcov.smle}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{smle}(object, ...)
}
\arguments{
\item{object}{
    a fitted model object, typically.  Sometimes also a
    \code{\link{summary}()} object of such a fitted model.
  }

\item{...}{
    additional arguments for method functions.  For the
    \code{\link{glm}} method this can be used to pass a
    \code{dispersion} parameter.
  }
}
\value{
A matrix of the estimated covariances between the parameter estimates
  in the linear or non-linear predictor of the model.  This should have
  row and column names corresponding to the parameter names given by the
  \code{\link{coef}} method.
}
\description{
Returns the variance-covariance matrix of the main parameters of
  a fitted model object.
}
\keyword{internal}
